Code style correction

+++ ImageLibrary/EditableImage.cs
@@ -1,13 +1,10 @@
 ﻿using System;
 using System.ComponentModel;
 using System.Drawing;
-using System.Drawing.Imaging;
 using System.IO;
 using System.Runtime.CompilerServices;
-using AForge.Imaging.Filters;
 using ImageLibrary.Annotations;
 using ImageLibrary.Filters;
-using IFilter = ImageLibrary.Filters.IFilter;
 
 namespace ImageLibrary
 {
@@ -16,7 +13,7 @@ namespace ImageLibrary
         private Bitmap _sourceImage;
         private Bitmap _imageView;
         private String _path;
-        private ComplementFilter filter;
+        private ComplementFilter _filter;
 
         public Bitmap ImageView
         {
@@ -43,28 +40,28 @@ namespace ImageLibrary
             _sourceImage = new Bitmap(path);
             _path = path;
             ImageView = _sourceImage;
-            filter = new ComplementFilter();
+            _filter = new ComplementFilter();
         }
 
         public void ChangeBrightness(int brightness)
         {
             IFilter f = new BrightnessFilter(brightness);
-            filter.AddOrUpdateFilter(f);
-            ImageView = filter.Apply(_sourceImage);
+            _filter.AddOrUpdateFilter(f);
+            ImageView = _filter.Apply(_sourceImage);
         }
 
         public void ChangeContrast(int contrast)
         {
             IFilter f = new ContrastFilter(contrast);
-            filter.AddOrUpdateFilter(f);
-            ImageView = filter.Apply(_sourceImage);
+            _filter.AddOrUpdateFilter(f);
+            ImageView = _filter.Apply(_sourceImage);
         }
 
         public void ChangeSaturation(int saturation)
         {
             IFilter f = new SaturationFilter(saturation);
-            filter.AddOrUpdateFilter(f);
-            ImageView = filter.Apply(_sourceImage);
+            _filter.AddOrUpdateFilter(f);
+            ImageView = _filter.Apply(_sourceImage);
         }
 
         public bool IsChanged()
@@ -75,26 +72,32 @@ namespace ImageLibrary
         public void Save()
         {
             var extension = Path.GetExtension(_path);
-            extension = extension.ToLower();
-            _sourceImage.Dispose();
-            File.Delete(_path);
-            ImageView.Save(_path, StringToImageExtensionConverter.Convert(extension));
+            if (extension != null)
+            {
+                extension = extension.ToLower();
+                _sourceImage.Dispose();
+                File.Delete(_path);
+                ImageView.Save(_path, StringToImageExtensionConverter.Convert(extension));
+            }
             _sourceImage = ImageView;
         }
 
         public void Save(String path)
         {
             var extension = Path.GetExtension(path);
-            extension = extension.ToLower();
-            ImageView.Save(path, StringToImageExtensionConverter.Convert(extension));
+            if (extension != null)
+            {
+                extension = extension.ToLower();
+                ImageView.Save(path, StringToImageExtensionConverter.Convert(extension));
+            }
             _sourceImage = ImageView;
         }
 
         public void ChangeColor(int red, int green, int blue)
         {
             IFilter f = new ColorFilter(red, green, blue);
-            filter.AddOrUpdateFilter(f);
-            ImageView = filter.Apply(_sourceImage);
+            _filter.AddOrUpdateFilter(f);
+            ImageView = _filter.Apply(_sourceImage);
         }
     }
 }
